testi user_group fix

+++ DSL/Resql/ljvis/POST/user_group/get_user_group_users.sql
@@ -72,8 +72,8 @@ SELECT
 FROM latest l
 WHERE
     (
-    :user_group_id = 0
-        OR :user_group_id = ANY(l.user_groups)
+    :user_group_id::BIGINT = 0
+        OR :user_group_id::BIGINT = ANY(l.user_groups)
         )
     AND
     (
